Implement fixed priority queue size
The simulation now uses a priority queue (heapdict) that allows for
changing the priority so the only necessary handlers are one for each
particle to handler wall collisions and one for each pair of particles
to handler interparticular collisions.

Performance is still lacking for 100 particles, but it is better than
the time-driven naive simulation performance.